requirements.txt
# Dependências principais
crewai>=0.28.0
langchain>=0.0.335
openai>=1.3.0

# Clientes de banco de dados
psycopg2-binary>=2.9.9
redis>=5.0.1
qdrant-client>=1.6.0

# Ferramentas HTTP
requests>=2.31.0
aiohttp>=3.9.1
websockets>=12.0

# Processamento de dados
numpy>=1.26.0
pandas>=2.1.3
pydantic>=2.5.2

# Logging e monitoramento
prometheus-client>=0.17.1
python-json-logger>=2.0.7

# Utilitários
python-dotenv>=1.0.0
tenacity>=8.2.3
tiktoken>=0.5.1
